FSPRA-1188 AES-GCM update datatype

+++ lib/include/tinycrypt/gcm_mode.h
@@ -102,7 +102,7 @@ int tc_gcm_encryption_init(const TCAesKeySched_t sched, uint8_t * iv, uint8_t * 
  * @param length IN -- intput length
  */
 int tc_gcm_encryption_update(const TCAesKeySched_t sched, const uint8_t * input, uint8_t * output,
-                             uint8_t length);
+                             uint32_t length);
 
 /**
  * @brief calculates TAG for GCM mode encryption procedure
@@ -123,8 +123,8 @@ int tc_gcm_encryption_update(const TCAesKeySched_t sched, const uint8_t * input,
  */
 int tc_gcm_encryption_final(const TCAesKeySched_t sched,
                             uint8_t             * input,
-                            uint8_t               input_len,
-                            uint8_t               aad_len,
+                            uint32_t              input_len,
+                            uint32_t              aad_len,
                             uint8_t             * output,
                             uint8_t             * tag);
 
@@ -157,7 +157,7 @@ int tc_gcm_decryption_init(const TCAesKeySched_t sched, uint8_t * iv, uint8_t * 
  * @param length IN -- intput length
  */
 int tc_gcm_decryption_update(const TCAesKeySched_t sched, const uint8_t * input, uint8_t * output,
-                             uint8_t length);
+                             uint32_t length);
 
 /**
  * @brief calculates TAG for GCM mode decryption procedure
@@ -181,8 +181,8 @@ int tc_gcm_decryption_update(const TCAesKeySched_t sched, const uint8_t * input,
 int tc_gcm_decryption_final(const TCAesKeySched_t sched,
                             uint8_t             * input,
                             uint8_t             * tag,
-                            uint8_t               aad_len,
-                            uint8_t               input_len,
+                            uint32_t              aad_len,
+                            uint32_t              input_len,
                             uint8_t               tag_len,
                             uint8_t             * output);
